zzform upload: zz_unlink_cleanup() now returns true if file does not exist at first
git-svn-id: file:///Repos/zzform/trunk@1121 98be6e6d-ea4f-4117-9388-ed847deb6ed9

+++ zzform/upload.inc.php
@@ -2236,6 +2236,7 @@ function zz_upload_get_typelist($filename, $type = 'Filetype', $optional = false
  */
 function zz_unlink_cleanup($file) {
 	$full_path = realpath($file);
+	if (!$full_path) return true;
 	$dir = dirname($full_path);
 	$success = unlink($full_path);
